Fix miscompilation of expressions with parentheses.
A bug in the LSCP_EMIT_ASSEMBLY pass for LLScriptParenthesis caused the type of the parenthesized expression to be lost. It assigned the type when it shouldn't, causing miscompilation of e.g. 1 + ([]). Fixed by letting it be handled by the default case, which already does the right thing (ignore the parentheses).